+++ packages/amplify-e2e-tests/src/__tests__/api_1.test.ts
@@ -4,7 +4,7 @@ import {
   deleteProject,
   initFlutterProjectWithProfile,
   initJSProjectWithProfile,
-  addApiWithSchema,
+  addApiWithoutSchema,
   updateApiSchema,
   updateApiWithMultiAuth,
   createNewProjectDir,
@@ -22,6 +22,7 @@ import {
   getBackendAmplifyMeta,
   amplifyPushUpdateForDependentModel,
   amplifyPushForce,
+  createRandomName,
 } from 'amplify-e2e-core';
 import path from 'path';
 import { existsSync } from 'fs';
@@ -30,8 +31,10 @@ import _ from 'lodash';
 
 describe('amplify add api (GraphQL)', () => {
   let projRoot: string;
+  let projFolderName: string;
   beforeEach(async () => {
-    projRoot = await createNewProjectDir('graphql-api');
+    projFolderName = 'graphqlapi';
+    projRoot = await createNewProjectDir(projFolderName);
   });
 
   afterEach(async () => {
@@ -44,8 +47,10 @@ describe('amplify add api (GraphQL)', () => {
 
   it('init a project and add the simple_model api', async () => {
     const envName = 'devtest';
-    await initJSProjectWithProfile(projRoot, { name: 'simplemodel', envName });
-    await addApiWithSchema(projRoot, 'simple_model.graphql');
+    const projName = 'simplemodel';
+    await initJSProjectWithProfile(projRoot, { name: projName, envName });
+    await addApiWithoutSchema(projRoot);
+    await updateApiSchema(projRoot, projName, 'simple_model.graphql');
     await amplifyPush(projRoot);
 
     const meta = getProjectMeta(projRoot);
@@ -74,8 +79,10 @@ describe('amplify add api (GraphQL)', () => {
 
   it('init a project then add and remove api', async () => {
     const envName = 'devtest';
-    await initIosProjectWithProfile(projRoot, { name: 'simplemodel', envName });
-    await addApiWithSchema(projRoot, 'simple_model.graphql');
+    const projName = 'simplemodel';
+    await initIosProjectWithProfile(projRoot, { name: projName, envName });
+    await addApiWithoutSchema(projRoot);
+    await updateApiSchema(projRoot, projName, 'simple_model.graphql');
     await amplifyPush(projRoot);
 
     let meta = getProjectMeta(projRoot);
@@ -98,8 +105,10 @@ describe('amplify add api (GraphQL)', () => {
 
   it('init a Flutter project and add the simple_model api', async () => {
     const envName = 'devtest';
-    await initFlutterProjectWithProfile(projRoot, { name: 'simplemodel', envName });
-    await addApiWithSchema(projRoot, 'simple_model.graphql');
+    const projName = 'simplemodel';
+    await initFlutterProjectWithProfile(projRoot, { name: projName, envName });
+    await addApiWithoutSchema(projRoot);
+    await updateApiSchema(projRoot, projName, 'simple_model.graphql');
     await amplifyPushWithoutCodegen(projRoot);
 
     const meta = getProjectMeta(projRoot);
@@ -132,7 +141,8 @@ describe('amplify add api (GraphQL)', () => {
     const initialSchema = 'initial_key_blog.graphql';
     const nextSchema = 'next_key_blog.graphql';
     await initJSProjectWithProfile(projRoot, { name: projectName });
-    await addApiWithSchema(projRoot, initialSchema);
+    await addApiWithoutSchema(projRoot);
+    await updateApiSchema(projRoot, projectName, initialSchema);
     await amplifyPush(projRoot);
     updateApiSchema(projRoot, projectName, nextSchema);
     await amplifyPushUpdate(projRoot);
@@ -145,13 +155,15 @@ describe('amplify add api (GraphQL)', () => {
   });
 
   it('init a project and add the simple_model api with multiple authorization providers', async () => {
-    await initJSProjectWithProfile(projRoot, { name: 'simplemodelmultiauth' });
-    await addApiWithSchema(projRoot, 'simple_model.graphql');
+    const appName = createRandomName();
+    await initJSProjectWithProfile(projRoot, { name: appName });
+    await addApiWithoutSchema(projRoot);
+    await updateApiSchema(projRoot, appName, 'simple_model.graphql');
     await updateApiWithMultiAuth(projRoot, {});
     await amplifyPush(projRoot);
 
     const meta = getProjectMeta(projRoot);
-    const { output } = meta.api.simplemodelmultiauth;
+    const { output } = meta.api[appName];
     const { GraphQLAPIIdOutput, GraphQLAPIEndpointOutput, GraphQLAPIKeyOutput } = output;
     const { graphqlApi } = await getAppSyncApi(GraphQLAPIIdOutput, meta.providers.awscloudformation.Region);
 
@@ -189,7 +201,8 @@ describe('amplify add api (GraphQL)', () => {
   it('init a project and add the simple_model api, match transformer version to current version', async () => {
     const name = `simplemodelv${TRANSFORM_CURRENT_VERSION}`;
     await initJSProjectWithProfile(projRoot, { name });
-    await addApiWithSchema(projRoot, 'simple_model.graphql');
+    await addApiWithoutSchema(projRoot);
+    await updateApiSchema(projRoot, name, 'simple_model.graphql');
     await amplifyPush(projRoot);
 
     const meta = getProjectMeta(projRoot);
@@ -217,7 +230,8 @@ describe('amplify add api (GraphQL)', () => {
     const initialSchema = 'two-model-schema.graphql';
     const fnName = `integtestfn${random}`;
     await initJSProjectWithProfile(projRoot, { name: projectName });
-    await addApiWithSchema(projRoot, initialSchema);
+    await addApiWithoutSchema(projRoot);
+    await updateApiSchema(projRoot, projectName, initialSchema);
     await addFunction(
       projRoot,
       {
@@ -264,7 +278,8 @@ describe('amplify add api (GraphQL)', () => {
   it('api force push with no changes', async () => {
     const projectName = `apinochange`;
     await initJSProjectWithProfile(projRoot, { name: projectName });
-    await addApiWithSchema(projRoot, 'two-model-schema.graphql');
+    await addApiWithoutSchema(projRoot);
+    await updateApiSchema(projRoot, projectName, 'two-model-schema.graphql');
     await amplifyPush(projRoot);
     let meta = getBackendAmplifyMeta(projRoot);
     const { lastPushDirHash: beforeDirHash } = meta.api[projectName];