+++ packages/amplify-e2e-tests/src/__tests__/schema-iterative-update-1.test.ts
@@ -4,19 +4,24 @@ import {
   initJSProjectWithProfile,
   deleteProject,
   deleteProjectDir,
-  addApiWithSchema,
+  addApiWithoutSchema,
   addFeatureFlag,
   amplifyPush,
   updateApiSchema,
   amplifyPushUpdate,
+  createRandomName,
 } from 'amplify-e2e-core';
 
 describe('Schema iterative update - rename @key', () => {
   let projectDir: string;
+  let appName: string;
 
   beforeAll(async () => {
+    appName = createRandomName();
     projectDir = await createNewProjectDir('schemaIterative');
-    await initJSProjectWithProfile(projectDir, {});
+    await initJSProjectWithProfile(projectDir, {
+      name: appName,
+    });
 
     addFeatureFlag(projectDir, 'graphqltransformer', 'enableiterativegsiupdates', true);
   });
@@ -25,14 +30,13 @@ describe('Schema iterative update - rename @key', () => {
     deleteProjectDir(projectDir);
   });
   it('should support changing gsi name', async () => {
-    const apiName = 'renamekey';
-
     const initialSchema = path.join('iterative-push', 'change-model-name', 'initial-schema.graphql');
-    await addApiWithSchema(projectDir, initialSchema, { apiName, apiKeyExpirationDays: 7 });
+    await addApiWithoutSchema(projectDir, { apiKeyExpirationDays: 7 });
+    await updateApiSchema(projectDir, appName, initialSchema);
     await amplifyPush(projectDir);
 
     const finalSchema = path.join('iterative-push', 'change-model-name', 'final-schema.graphql');
-    await updateApiSchema(projectDir, apiName, finalSchema);
+    await updateApiSchema(projectDir, appName, finalSchema);
     await amplifyPushUpdate(projectDir, undefined, undefined, true);
   });
 });